Fixed player damage bug
was shooting shotgun each time. Had no check for if shotgun was equipped

+++ Assets/Scripts/EnemyAI.cs
@@ -143,7 +143,7 @@ public class EnemyAI : MonoBehaviour, IDamage
         }
     }
 
-    public void takeDamage(float dmg)
+    public void takeDamage(int dmg)
     {
         HP -= dmg;
