+++ Assets/Scripts/IDamage.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IDamage 
 {
-    void takeDamage(float dmg);
+    void takeDamage(int dmg);
 
 }
